+++ pages/components/Editor.tsx
@@ -1,7 +1,7 @@
 import { SetStateAction, useEffect, useRef, useState } from "react";
 
+import { UseWordiablesContext } from "../../context/wordiablesContext";
 import styles from "../../styles/editor.module.scss";
-import { UseWordiablesContext } from "../context/wordiablesContext";
 import LiveEditor from "./LiveEditor";
 
 // import { stringToElement } from "../lib/editor";
@@ -23,8 +23,8 @@ export default function Editor (){
 					placeholder='Paste your cover letter here'
 					onChange={(e) => setText(e.target.value)}
 				/>
+				<LiveEditor>{text}</LiveEditor>
 			</div>
-			<LiveEditor>{text}</LiveEditor>
 		</section>
 	);
 }
