+++ pages/components/Wordiables.tsx
