+++ pages/components/logo.tsx
