+++ styles/editor.module.scss
@@ -16,17 +16,17 @@
 	border: 1px solid #000;
 }
 
+.heading {
+	margin: 0;
+	color: #525252;
+}
+
 .editor {
 	height: 100%;
 	width: 100%;
 	background: #fff;
 	overflow-y: auto;
 	padding: 1rem;
-}
-
-.heading {
-	margin: 0;
-	color: #525252;
 }
 
 .textEditor {
